add SalatRepository and refactoring
